+++ client/package.json
@@ -10,7 +10,7 @@
   },
   "scripts": {
     "test": "npm run test:karma && npm run test:legacy",
-    "test:legacy": "NODE_ENV=test mocha --recursive test/setup.js test/* --compilers jsx:babel-core/register js:babel-core/register",
+    "test:legacy": "NODE_ENV=test mocha --recursive test/node/setup.js test/node/* --compilers jsx:babel-core/register js:babel-core/register",
     "test:karma": "NODE_ENV=test karma start",
     "test:karma:watch": "npm run test:karma -- --watch",
     "test:legacy:watch": "npm run test:legacy -- --watch",
